hypertable: no more autodetection :(
Got autodetection for > 99% of the time, but there will
always be dirty or lossy data that will prevent it from
guessing correctly all the time.  Instead, it is now assumed
that the "search" table contains a value with the
expected rate.  We populate this using a storage-schema.conf
file from our 'carbons'